+++ src/rscomplex.c
@@ -1,0 +1,42 @@
+#include "rsbench.h"
+RSComplex c_add( RSComplex A, RSComplex B)
+{
+	RSComplex C;
+	C.r = A.r + B.r;
+	C.i = A.i + B.i;
+	return C;
+}
+RSComplex c_sub( RSComplex A, RSComplex B)
+{
+	RSComplex C;
+	C.r = A.r - B.r;
+	C.i = A.i - B.i;
+	return C;
+}
+RSComplex c_mul( RSComplex A, RSComplex B)
+{
+	double a = A.r;
+	double b = A.i;
+	double c = B.r;
+	double d = B.i;
+	RSComplex C;
+	C.r = (a*c) - (b*d);
+	C.i = (a*d) + (b*c);
+	return C;
+}
+RSComplex c_div( RSComplex A, RSComplex B)
+{
+	double a = A.r;
+	double b = A.i;
+	double c = B.r;
+	double d = B.i;
+	RSComplex C;
+	double denom = c*c + d*d;
+	C.r = ( (a*c) + (b*d) ) / denom;
+	C.i = ( (b*c) - (a*d) ) / denom;
+	return C;
+}
+double c_abs( RSComplex A)
+{
+	return sqrt(A.r*A.r + A.i * A.i);
+}
